designer/scripts/requirements.txt
supabase>=2.0.0
requests
beautifulsoup4
python-dotenv
schedule
